ml_api/requirements.txt
fastapi
uvicorn[standard]
joblib
numpy
google-cloud-firestore
scikit-learn
pydantic
